# Conflicts:
#	.idea/workspace.xml
#	cmake-build-debug/CMakeCache.txt
#	cmake-build-debug/CMakeFiles/CMakeDirectoryInformation.cmake
#	cmake-build-debug/CMakeFiles/CMakeOutput.log
#	cmake-build-debug/CMakeFiles/Makefile.cmake
#	cmake-build-debug/CMakeFiles/Makefile2
#	cmake-build-debug/CMakeFiles/Projeto1.dir/DependInfo.cmake
#	cmake-build-debug/CMakeFiles/Projeto1.dir/build.make
#	cmake-build-debug/CMakeFiles/Projeto1.dir/compiler_depend.make
#	cmake-build-debug/CMakeFiles/Projeto1.dir/flags.make
#	cmake-build-debug/CMakeFiles/Projeto1.dir/link.txt
#	cmake-build-debug/CMakeFiles/TargetDirectories.txt
#	cmake-build-debug/CMakeFiles/clion-environment.txt
#	cmake-build-debug/Makefile
#	cmake-build-debug/Projeto1.cbp
#	cmake-build-debug/Testing/Temporary/LastTest.log
#	cmake-build-debug/cmake_install.cmake
#	src/NetworkManager.cpp

+++ src/Network.cpp
@@ -7,7 +7,7 @@
 Network::Network(std::string stationA, std::string stationB)
     : stationA_(stationA), stationB_(stationB) {}
 
-Network::Network(std::string stationA, std::string stationB, int capacity, std::string service)
+Network::Network(std::string stationA, std::string stationB, double capacity, std::string service)
     : stationA_(stationA),
     stationB_(stationB),
     capacity_(capacity),
@@ -19,7 +19,7 @@ const string Network::getStationA() const {
 const string Network::getStationB() const {
     return stationB_;
 }
-const int Network::getCapacity() const {
+const double Network::getCapacity() const {
     return capacity_;
 }
 const string Network::getService() const {
@@ -32,7 +32,7 @@ void Network::setStationA(const std::string stationA) {
 void Network::setStationB(const std::string stationB) {
     stationB_ = stationB;
 }
-void Network::setCapacity(const int capacity) {
+void Network::setCapacity(const double capacity) {
     capacity_ = capacity;
 }
 void Network::setService(const std::string service) {
